+++ lib/features/accounting/domain/repositories/chart_of_accounts_repository.dart
@@ -2,7 +2,7 @@ import '../entities/account_node.dart';
 
 abstract interface class ChartOfAccountsRepository {
   Future<List<AccountNode>> getAccounts();
+  Future<String> previewNextCode(AccountLevel level, {String? parentId});
   Future<AccountNode> createAccount(AccountNode account);
   Future<AccountNode> updateAccount(AccountNode account);
 }
-